Existing network support + API Changes
This alters the API to take the notion of options; a pattern used in
other libraries such as gRPC. Now, instead of passing a network name in
your New() calls, you will pass one of WithNewNetwork() or
WithExistingNetwork().

WithNewNetwork() will create a new network, and WithExistingNetwork will
take a network id of a previously launched composition (which you can
get with (*Composer).GetNetworkID()).

Neat, eh? Check the TestWithExistingNetwork test for more information.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>